+++ R/burgle.R
@@ -28,8 +28,20 @@ burgle.lm <- function(object, ...){
   cov <- stats::vcov(object)
   rss <- sum(object$residuals ^2)/object$df.residual
   xlevels <- object$xlevels
-  # formula <- stats::reformulate(as.character(attr(object$terms, "predvars"))[-c(1:2)])
   formula <- (as.character(attr(object$terms, "predvars"))[-c(1:2)])
+
+  ## interactions
+  tlo <- attr(object$terms, "order")
+  if(any(tlo >1)){
+    tl <- attr(object$terms, "term.labels")
+    tl0 <- tl[which(tlo <=1)]
+    tli <- tl[which(tlo >1)]
+    tli2 <- strsplit(tli, "(?<!:)(:)(?!:)", perl = T)
+    formula <- c(formula, sapply(tli2, function(x) make_ints(x, o_form = formula, tl0 = tl0)))
+  }
+
+  if(attr(object$terms, "intercept") == 0) formula <- c(formula, "-1")
+  ##
 
   l <- list("coef" = coef,
             "cov" = cov,
@@ -56,8 +68,18 @@ burgle.glm <- function(object, ...){
 
   xlevels <- object$xlevels
 
-  # formula <- stats::reformulate(as.character(attr(object$terms, "predvars"))[-c(1:2)])
   formula <- (as.character(attr(object$terms, "predvars"))[-c(1:2)])
+
+  ## interactions
+  tlo <- attr(object$terms, "order")
+  if(any(tlo >1)){
+    tl <- attr(object$terms, "term.labels")
+    tl0 <- tl[which(tlo <=1)]
+    tli <- tl[which(tlo >1)]
+    tli2 <- strsplit(tli, "(?<!:)(:)(?!:)", perl = T)
+    formula <- c(formula, sapply(tli2, function(x) make_ints(x, o_form = formula, tl0 = tl0)))
+  }
+  if(attr(object$terms, "intercept") == 0) formula <- c(formula, "-1")
 
   family <- object$family$family