+++ NEWS.md
@@ -1,3 +1,14 @@
+# Changes in version 0.1.1
+
+## New features
+
+* Now allows zero-intercept models and will incorporate interaction terms
+* burgle methods now exist for nnet::multinom models
+
+## Minor improvements
+
+* Set the default to original = TRUE for all predict methods
+
 # burgle 0.1.0
 
 * Initial CRAN submission.